requirements.txt
lightning==2.2.1
matplotlib==3.7.1
numba==0.57.0
numpy==1.24.3
pandas==1.5.3
pytorch-lightning==2.2.1
PyYAML==6.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
tifffile==2023.9.26
timm==0.9.2
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torcheval==0.0.7
torchmetrics==1.2.1
torchvision==0.15.2+cu118
tqdm==4.65.0
transformers==4.29.2
ipykernel==6.23.1
ipython==8.13.2
ipywidgets==8.0.6
jupyterlab==4.0.0